requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
idna==3.10
psycopg==3.2.9
psycopg-binary==3.2.9
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.3
typing_extensions==4.13.2
urllib3==2.4.0
